BRO JUST DO THIS OML

+++ Playlist.hpp
@@ -24,13 +24,12 @@ class Playlist {
 
 	double getTotalMinutes() const; // returns the total play time.
 
-	int averageSongLength() const; // returns the average song length in minutes (totalMinutes_ / numSongs_). If numSongs_ is 0, return 0.
+	double averageSongLength() const; // returns the average song length in minutes (totalMinutes_ / numSongs_). If numSongs_ is 0, return 0.
 
 	Playlist operator+(const Playlist& other) const; //Returns a new Playlist whose numSongs_ and totalMinutes_ are the sum of the two operands.
 
-	/*
 	Playlist& operator+=(const Playlist& other); //Modifies the current playlist by adding in another playlist’s songs and time.
-	*/
+
 };
